requirements.txt
aws-cdk-lib==2.55.1
aws-cdk.aws-glue-alpha==2.55.1a0
constructs>=10.0.0,<11.0.0
